led_light_sensor_values.py
This program extends on the original Light_Sensor program to detect different levels of light in a room or given area. And triggers 3 LEDS depending on if the lighting is Dim, Bright, or Dark. 

This can be extended to then trigger different actions, sensors, etc based on the value of how much light is being received.