Update LED & Light Values
Updated LightSensor.py to led_light_sensor_values.py

The reason for this change is that I added onto the program to include using the count value (which adjusts depending on the light in the room). The LED's use the count/light value to blink depending on whether the room or area is in Bright, Dim, or Dark/No Lighting.

I'll be including a diagram for this eventually as well.